requirements.txt
# MEGABOT Dependencies

# Data Collection
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
requests>=2.31.0

# AI APIs
anthropic>=0.39.0  # Claude
google-generativeai>=0.8.3  # Gemini

# Dashboard
streamlit>=1.40.0
plotly>=5.18.0

# Configuration
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
